refactor(server): move backend source files into src directory
