+++ server/src/routes/friends.js
