+++ server/src/services/friendService.js
